+++ packages/design-system/src/components/Slider/index.ts
@@ -1,0 +1,8 @@
+export { sliderVariants } from './classes';
+export { Slider, type SliderProps } from './Slider';
+export { SliderControl, type SliderControlProps } from './SliderControl';
+export { SliderInput, type SliderInputProps } from './SliderInput';
+export { SliderMarks, type SliderMarksProps } from './SliderMarks';
+export { SliderThumb, type SliderThumbProps } from './SliderThumb';
+export { SliderValue, type SliderValueProps } from './SliderValue';
+export type { SliderMark } from './types';