+++ packages/design-system/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.70.0 (2026-07-02)
+
+* feat(simple-charts): add HorizontalBar chart (#198) ([043a29b](https://github.com/wallarm/design-system/commit/043a29b)), closes [#198](https://github.com/wallarm/design-system/issues/198)
+
+## 0.69.0 (2026-07-01)
+
+* feat(design-system): add Slider component (#192) ([7978fb4](https://github.com/wallarm/design-system/commit/7978fb4)), closes [#192](https://github.com/wallarm/design-system/issues/192)
+
+## <small>0.68.2 (2026-06-29)</small>
+
+* fix(filter-input): AS-1171 cross-field label borrowing (#191) ([0a06b8e](https://github.com/wallarm/design-system/commit/0a06b8e)), closes [#191](https://github.com/wallarm/design-system/issues/191)
+* docs(filter-input): mirror new Attacks schema in Composition Default story (#190) ([77fec9d](https://github.com/wallarm/design-system/commit/77fec9d)), closes [#190](https://github.com/wallarm/design-system/issues/190)
+
 ## <small>0.68.1 (2026-06-26)</small>
 
 * fix(filter-input): AS-1171 keep type-conforming values raw in chips (#189) ([9bb0154](https://github.com/wallarm/design-system/commit/9bb0154)), closes [#189](https://github.com/wallarm/design-system/issues/189)